vvefxbot/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytz>=2023.3
gspread>=5.10.0
oauth2client>=4.1.3
pyTelegramBotAPI>=4.14.0
aiohttp>=3.9.0
schedule>=1.2.0
